+++ src/framework/DOCExecutable.ts
@@ -14,6 +14,7 @@ import { XmlParser, XmlTag, XmlTagProcessor } from '../util/XmlParser';
 import { assert, debugAssert, debugBreak } from '../util/DevelopmentHelper';
 import { TestResultBuilder } from '../TestResultBuilder';
 import { TestItemParent } from '../TestItemManager';
+import { SubTestTree } from '../AbstractTest';
 
 export class DOCExecutable extends AbstractExecutable {
   public constructor(shared: WorkspaceShared, execInfo: RunnableProperties, docVersion: Version | undefined) {
@@ -318,12 +319,13 @@ abstract class TagProcessorBase implements XmlTagProcessor {
     public readonly builder: TestResultBuilder,
     protected readonly shared: WorkspaceShared,
     protected readonly caseData: CaseData,
+    protected readonly subCases: SubTestTree,
   ) {}
 
   public onopentag(tag: XmlTag): void | XmlTagProcessor | Promise<void | XmlTagProcessor> {
     const procCreator = TagProcessorBase.openTagProcessorMap.get(tag.name);
     if (procCreator) {
-      return procCreator(tag, this.builder, this.caseData, this.shared);
+      return procCreator(tag, this.builder, this.caseData, this.shared, this.subCases);
     } else if (procCreator === null) {
       // known tag, do nothing
     } else {
@@ -371,6 +373,7 @@ abstract class TagProcessorBase implements XmlTagProcessor {
         builder: TestResultBuilder,
         caseData: CaseData,
         shared: WorkspaceShared,
+        subCases: SubTestTree,
       ) => void | XmlTagProcessor | Promise<void | XmlTagProcessor>)
   > = new Map([
     [
@@ -380,9 +383,10 @@ abstract class TagProcessorBase implements XmlTagProcessor {
         builder: TestResultBuilder,
         caseData: CaseData,
         shared: WorkspaceShared,
+        subCases: SubTestTree,
       ): void | XmlTagProcessor | Promise<XmlTagProcessor> => {
         // if name is missing we don't create subcase for it
-        if (tag.attribs.name) return SubCaseProcessor.create(shared, builder, tag.attribs, caseData);
+        if (tag.attribs.name) return SubCaseProcessor.create(shared, builder, tag.attribs, caseData, subCases);
       },
     ],
     [
@@ -438,7 +442,7 @@ class TestCaseTagProcessor extends TagProcessorBase {
     private readonly attribs: Record<string, string>,
     _option: Option,
   ) {
-    super(builder, shared, {});
+    super(builder, shared, {}, new Map());
   }
 
   public async begin(): Promise<void> {
@@ -474,6 +478,7 @@ class TestCaseTagProcessor extends TagProcessorBase {
       }
 
       this.builder[result]();
+      TODO: this.builder.test.removeMissingSubTests(this.subCases);
       this.builder.build();
     } else {
       return super.onopentag(tag);
@@ -489,6 +494,7 @@ class SubCaseProcessor extends TagProcessorBase {
     testBuilder: TestResultBuilder,
     attribs: Record<string, string>,
     parentCaseData: CaseData,
+    subCases: SubTestTree,
   ): Promise<SubCaseProcessor> {
     if (typeof attribs.name !== 'string' || !attribs.name) throw Error('Section must have name attribute');
 
@@ -501,13 +507,24 @@ class SubCaseProcessor extends TagProcessorBase {
     const subTest = await testBuilder.test.getOrCreateSubTest(attribs.name, label, attribs.filename, attribs.line);
     const subTestBuilder = testBuilder.createSubTestBuilder(subTest);
 
-    return new SubCaseProcessor(shared, subTestBuilder, parentCaseData);
+    let subSubCases = subCases.get(attribs.name);
+    if (subSubCases === undefined) {
+      subSubCases = new Map();
+      subCases.set(attribs.name, subSubCases);
+    }
+
+    return new SubCaseProcessor(shared, subTestBuilder, parentCaseData, subSubCases);
   }
 
-  private constructor(shared: WorkspaceShared, testBuilder: TestResultBuilder, parentCaseData: CaseData) {
+  private constructor(
+    shared: WorkspaceShared,
+    testBuilder: TestResultBuilder,
+    parentCaseData: CaseData,
+    subCases: SubTestTree,
+  ) {
     testBuilder.started();
 
-    super(testBuilder, shared, parentCaseData);
+    super(testBuilder, shared, parentCaseData, subCases);
   }
 
   //doctest does not provide result for sub-cases, no point to build