+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- `envFile` now supports `.env` format too. (determines from the end of the filename)
+
 ## [4.0.7] - 2021-11-27
 
 ### Fixed