+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [4.0.8]
+
 ### Added
 
 - `envFile` now supports `.env` format too. (determines from the end of the filename)
+- removing missing SubCases / Sections
 
 ## [4.0.7] - 2021-11-27